requirements.txt
annotated-types==0.7.0
anyio==4.6.0
biopython==1.84
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
interchange==2021.0.4
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
monotonic==1.6
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pansi==2020.7.3
pika==1.3.2
pluggy==1.5.0
py2neo==2021.2.4
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
pytest-mock==3.14.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchdog==5.0.3
watchfiles==0.24.0
websockets==13.1
